+++ app/components/status_tags/admin_participant_status_tag.rb
@@ -31,7 +31,7 @@ module StatusTags
     end
 
     def record_state
-      @record_state ||= DetermineTrainingRecordState.call(participant_profile:, induction_record:, school:)&.record_state || :no_longer_involved
+      @record_state ||= DetermineTrainingRecordStateLegacy.call(participant_profile:, induction_record:, school:)&.record_state || :no_longer_involved
     end
   end
 end
