+++ app/services/appropriate_bodies/induction_records_query.rb
@@ -16,6 +16,21 @@ module AppropriateBodies
 
       InductionRecord.distinct
         .joins("JOIN (#{join.to_sql}) AS latest_induction_records ON latest_induction_records.latest_id = induction_records.id")
+        .joins("LEFT JOIN email_associations ON email_associations.object_id = induction_records.participant_profile_id")
+        .joins("LEFT JOIN emails ON emails.id = email_associations.email_id AND 'request_for_details' = ANY (tags)")
+        .select(
+          "induction_records.*",
+          "emails.status AS transient_latest_request_for_details_status",
+          "EXISTS (
+            SELECT 1 FROM induction_records as mir
+              WHERE mir.mentor_profile_id = induction_records.participant_profile_id
+            ) AS transient_mentees",
+          "EXISTS (
+            SELECT 1 FROM induction_records as cmir
+              WHERE cmir.mentor_profile_id = induction_records.participant_profile_id
+                AND (cmir.induction_status = 'active' OR cmir.induction_status = 'leaving')
+            ) AS transient_current_mentees",
+        )
     end
 
   private
