+++ app/views/admin/participants/statuses/show.html.erb
@@ -7,7 +7,7 @@
 
 <%= render partial: "admin/participants/nav" %>
 
-<% states = DetermineTrainingRecordState.call(participant_profile: @participant_profile, school: @participant_presenter.school) %>
+<% states = DetermineTrainingRecordStateLegacy.call(participant_profile: @participant_profile, school: @participant_presenter.school) %>
 
 <h2 class="govuk-heading-m">School</h2>
 <% if @participant_presenter.school.blank? %>
@@ -91,7 +91,7 @@
 
     sl.with_row do |row|
       row.with_key(text: "Training record state")
-      row.with_value(text: render(StatusTags::AppropriateBodyParticipantStatusTag.new(participant_profile: @participant_presenter.participant_profile, appropriate_body: @participant_presenter.school_cohort&.appropriate_body)))
+      row.with_value(text: render(StatusTags::AppropriateBodyParticipantStatusTag.new(@appropriate_body_training_record_states[@participant_presenter.participant_profile.id])))
     end
   end %>
 <% end %>
