+++ spec/components/appropriate_bodies/participants/table_row_spec.rb
@@ -5,8 +5,9 @@ RSpec.describe AppropriateBodies::Participants::TableRow, type: :component do
   let(:school) { create(:school) }
   let(:school_cohort) { create(:school_cohort, school:) }
   let!(:induction_record) { create(:induction_record, participant_profile:, induction_programme:, training_status: "withdrawn") }
+  let(:training_record_states) { DetermineTrainingRecordState.call(participant_profiles: participant_profile, induction_records: induction_record) }
 
-  let(:component) { described_class.new induction_record: }
+  let(:component) { described_class.new induction_record:, training_record_states: }
 
   subject { render_inline(component) }
