+++ spec/components/status_tags/admin_participant_status_tag_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe StatusTags::AdminParticipantStatusTag, type: :component do
   subject(:label) { render_inline component }
 
   context "The language file" do
-    DetermineTrainingRecordState::RECORD_STATES.each_key do |key|
+    TrainingRecordState::RECORD_STATES.each_key do |key|
       it "includes the record_state :#{key} as a language entry" do
         expect(I18n.t("status_tags.admin_participant_status").keys).to include key.to_sym
       end
@@ -22,7 +22,7 @@ RSpec.describe StatusTags::AdminParticipantStatusTag, type: :component do
     end
 
     it "includes :#{key} as a recognised record_state" do
-      expect(DetermineTrainingRecordState::RECORD_STATES.keys).to include key.to_s
+      expect(TrainingRecordState::RECORD_STATES.keys).to include key.to_s
     end
   end
 end
